:text: Renaming live-tests file
