+++ omisego-admin/src/liveTest/kotlin/co/omisego/omisego/live/GetWalletLiveTest.kt
@@ -1,4 +1,4 @@
-package co.omisego.omisego.features
+package co.omisego.omisego.live
 
 /*
  * OmiseGO
@@ -9,8 +9,9 @@ package co.omisego.omisego.features
 
 import co.omisego.omisego.LiveTest
 import co.omisego.omisego.model.params.LoginParams
-import co.omisego.omisego.model.transaction.list.TransactionListParams
+import co.omisego.omisego.model.params.WalletParams
 import org.amshove.kluent.shouldBe
+import org.amshove.kluent.shouldNotBe
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -19,7 +20,7 @@ import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
 @Config(sdk = [23])
-class ListTransactionTest : LiveTest() {
+class GetWalletLiveTest : LiveTest() {
     private val secret by lazy { loadSecretFile("secret.json") }
 
     @Before
@@ -34,30 +35,24 @@ class ListTransactionTest : LiveTest() {
     }
 
     @Test
-    fun `list transaction should be returned successfully`() {
-        val response = client.getTransactions(
-            TransactionListParams.create(
-                searchTerm = null
-            )
+    fun `get a user wallet should be returned associated user of the account`() {
+        val response = client.getWallet(
+            WalletParams(secret.getString("user_address"))
         ).execute()
 
         response.isSuccessful shouldBe true
-        response.body()?.data?.pagination?.perPage shouldBe 10
-        response.body()?.data?.pagination?.currentPage shouldBe 1
-        response.body()?.data?.data?.size shouldBe 10
+        response.body()?.data?.user shouldNotBe null
+        response.body()?.data?.account shouldBe null
     }
 
     @Test
-    fun `list transaction with a specific account should return transactions associated with the account`() {
-        val response = client.getTransactions(
-            TransactionListParams.create(
-                searchTerm = secret.getString("account_address")
-            )
+    fun `get an account wallet should be returned associated account of the account`() {
+        val response = client.getWallet(
+            WalletParams(secret.getString("account_address"))
         ).execute()
 
         response.isSuccessful shouldBe true
-        response.body()?.data?.data?.forEach {
-            (secret.getString("account_address") in arrayOf(it.from.address, it.to.address)) shouldBe true
-        }
+        response.body()?.data?.account shouldNotBe null
+        response.body()?.data?.user shouldBe null
     }
 }
